D2X-XL failed to load mines with more than 5461 (32767 / 6) segments.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@982 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/gamemine.c
@@ -1653,7 +1653,7 @@ for (i = startI; i < endI; i++)
 
 //------------------------------------------------------------------------------
 
-static short loadIdx = 0;
+static int loadIdx = 0;
 static int loadOp = 0;
 static CFILE *mineDataFile;
